Added test_adc5g.py and a shell script to freeze and compile it.
This produces a standalone binary which will be used to run tests
on the PowerPC of a ROACH or ROACH2. Currently however the test
script just prints something silly out. The binary must be compiled
on a ROACH or ROACH2 and you need 'python2.6-examples' installed.
To compile simply run 'freezeit.sh'.