Added a new module to adc5g that emulates corr.katcp_wrapper.FpgaClient.
This module has a dummy class that simply returns fake data for most
of the typical methods of FpgaClient. And more importantly it has a "local"
version of FpgaClient that runs on the PowerPC of a ROACH or ROACH2 and
interfaces with the FPGA over the OS file handles created by BORPH, i.e.
it does not need tcpborphserver.