Added the adc5g_test MDL and BOF files.
WARNING: as of now adc5g_test.mdl WILL NOT COMPILE with the present library
commit of sma-wideband/mlib_devel@b0825a5f. This is due to restrictions on
the MMCM in HIGH bandwidth mode and clocks going through the BUFR network.
A work-around for this is in the works.