[rww_tools] Made this module runnable from the command line.
This can now be run directly in two ways. To clear ADC core-to-core
parameters use it this way:

$ ./rww_tools.py clear roach2-01 [...] roach2-08

Where the last arguments are a list of ROACH2 hostnames with ADC5g's.

To update core-to-core offsets using the og_from_noise routine:

$ ./rww_tools.py update roach2-01 [...] roach2-08

This will run og_from_noise on each ZDOK and generate the optimal
core-to-core offsets to remove interleaving spurs. Note that the
og_from_noise routine needs write permission in whatever directory
you run it in.